common/bitutil: optimize ANDBytes
