+++ common/bitutil/bitutil.go
@@ -62,32 +62,7 @@ func safeXORBytes(dst, a, b []byte) int {
 // ANDBytes ands the bytes in a and b. The destination is assumed to have enough
 // space. Returns the number of bytes and'd.
 func ANDBytes(dst, a, b []byte) int {
-	if supportsUnaligned {
-		return fastANDBytes(dst, a, b)
-	}
-	return safeANDBytes(dst, a, b)
-}
-
-// fastANDBytes ands in bulk. It only works on architectures that support
-// unaligned read/writes.
-func fastANDBytes(dst, a, b []byte) int {
-	n := len(a)
-	if len(b) < n {
-		n = len(b)
-	}
-	w := n / wordSize
-	if w > 0 {
-		dw := *(*[]uintptr)(unsafe.Pointer(&dst))
-		aw := *(*[]uintptr)(unsafe.Pointer(&a))
-		bw := *(*[]uintptr)(unsafe.Pointer(&b))
-		for i := 0; i < w; i++ {
-			dw[i] = aw[i] & bw[i]
-		}
-	}
-	for i := n - n%wordSize; i < n; i++ {
-		dst[i] = a[i] & b[i]
-	}
-	return n
+	return andBytes(dst, a, b)
 }
 
 // safeANDBytes ands one by one. It works on all architectures, independent if
